+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ```
     * Removed billingAddress and userData form SDK request.
     * Imported a missing class for the admin controller.
+    * Force transaction type to Auth when Order total amount is 0.
 ```
 
 # 1.0.0